Consolidated getEditionData into EditionCard helper
